Replace FixtureContext with more fine-grained contexts.
