+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bldrs",
-  "version": "1.0.797",
+  "version": "1.0.800",
   "main": "src/index.jsx",
   "license": "MIT",
   "homepage": "https://github.com/bldrs-ai/Share",
