+++ src/Components/TableRow.fixture.jsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import Box from '@mui/material/Box'
-import FixtureContext from '../FixtureContext'
+import {ThemeCtx} from '../theme/Theme.fixture'
 import TableRow from './TableRow'
 
 
 export default (
-  <FixtureContext>
+  <ThemeCtx>
     <Box sx={{width: '400px'}}>
       <TableRow heading='hello' subtext='hello' inputType='select' options={['Option1', 'Option2', 'Option3']}/>
     </Box>
-  </FixtureContext>
+  </ThemeCtx>
 )
