+++ src/theme/Theme.test.jsx
@@ -1,20 +1,19 @@
 import {Themes, getSystemCurrentLightDark} from './Theme'
 
 
-// https://stackoverflow.com/questions/41885841/how-can-i-mock-the-javascript-window-object-using-jest
-let windowSpy
-
-beforeEach(() => {
-  windowSpy = jest.spyOn(window, 'window', 'get')
-})
-
-
-afterEach(() => {
-  windowSpy.mockRestore()
-})
-
-
 describe('Theme', () => {
+  // https://stackoverflow.com/questions/41885841/how-can-i-mock-the-javascript-window-object-using-jest
+  let windowSpy
+
+  beforeEach(() => {
+    windowSpy = jest.spyOn(window, 'window', 'get')
+  })
+
+
+  afterEach(() => {
+    windowSpy.mockRestore()
+  })
+
   test('getSystemCurrentLightDark is day when system is day', () => {
     windowSpy.mockImplementation(() => ({
       matchMedia: jest.fn((query) => {
